Now loading the composition API in a seperate file
This is to load the API early enough. If not loaded in a seperate file,
the saves store will throw an error saying you can't use the composition
API before it is loaded.

Imports happens first in main.ts. Importing App.vue causes all
components to be imported, which means a component using the saves store
gets imported. The saves store calls createStore() at the top level, a
function which uses the composition API. Since the composition API
hasn't yet been loaded (since we're still in the import stage), the API
throws an error.