Overhauled games and saves data structure
Games have been moved into the saves store and is now a part of the state.

Both games and saves have been moved to an array of objects structure instead of the previous plain object structure. This will allow re-ordering in the future. Since they still need to be identified by an ID, the ID has been moved to the object itself.

Bosses, of both games and saves, have been changed to an array structure, since they don't need to be identified by an ID. Also, the bosses of a game is now just an array of strings, since their only property was the name of the boss.

The application now has the concept of a selected save, which replaces passing a save ID in the router. This simplifies displaying the info of a save side-by-side the saves list.

The saves list now is now organized by game.

createSave() returns a boolean indicating whether or not creating a save failed.